fix(libp2p): Fixing Read func in kotlinInterface.

+++ LibGoVsKotlin/gomobile/gomobile.go
@@ -21,7 +21,7 @@ func Open(ID string) int {
 	return kotlinHandler.Open(ID)
 }
 
-func Read(buf []byte, sourcePos, destPos, readLength int, ID string) int {
+func Read(buf []byte, sourcePos, destPos, readLength int, ID string) int64 {
 	return kotlinHandler.Read(buf, sourcePos, destPos, readLength, ID)
 }
